Modified
Modified Error
- Store.js 에서 modules 오타 확인 후 수정 (moduels -> modules)
