+++ vue-todo/src/components/TodoFooter.vue
@@ -5,30 +5,28 @@
 </template>
 
 <script>
-import { mapMutations } from 'vuex';
+import { mapMutations } from 'vuex'
 
 export default {
-
   methods: {
     ...mapMutations({
       clearTodo: 'clearAllItems'
-    }),
+    })
   }
 }
 </script>
 
 <style scoped>
-  .clearAllContainer {
-    width: 8.5rem;
-    height: 50px;
-    line-height: 50px;
-    background-color: white;
-    border-radius: 5px;
-    margin: 0 auto;
-  }
-
-  .clearAllBtn {
-    color: #e20303;
-    display: block;
-  }
+.clearAllContainer {
+  width: 8.5rem;
+  height: 50px;
+  line-height: 50px;
+  background-color: white;
+  border-radius: 5px;
+  margin: 0 auto;
+}
+.clearAllBtn {
+  color: #e20303;
+  display: block;
+}
 </style>
