+++ vue-todo/src/main.js
@@ -1,9 +1,10 @@
 import Vue from 'vue'
 import App from './App.vue'
 import { store } from './store/store'
-Vue.config.productionTip = false
+
 
 new Vue({
+  el: '#app',
   store,
   render: h => h(App),
-}).$mount('#app')
+})
